Publish 1.6
Now covers all five of the standard maps, selected based on the filename of the .ark selected.
The loading cog now continues to show while loading species to show on the map. Previously it would only show during the conversion processes and there was a confusing pause before points appeared on the map.