Publish v1.12
Adds support for Lost Island.

Note: the coordinate grid on the ingame map is significantly different to the what a GPS device says. I've chosen to make Larkator show the correct position on the drawn map, so it won't match what the ingame hand-held map shows exactly, but will be a much better representation of location.

Also updated ArkSavegameToolkit with the hopes of seeing less savegame loading issues.